% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcgov.R
\name{hcgov_dataset}
\alias{hcgov_dataset}
\title{Get detailed info about a specific dataset}
\usage{
hcgov_dataset(id)
}
\arguments{
\item{id}{Dataset identifier (e.g. "e4rr-zk4i")}
}
\value{
tibble: identifier, title, description, modified, access_level,
publisher_name, contact_name, contact_email, distribution_count
}
\description{
Get detailed info about a specific dataset
}
